move everything into job_object
